+++ package.json
@@ -13,11 +13,12 @@
   "scripts": {
     "lint": "eslint --ext .js,.vue lib",
     "release": "standard-version && git push --follow-tags && npm publish",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "jest test"
   },
   "devDependencies": {
     "@nuxtjs/eslint-config": "^2.0.2",
     "eslint": "^6.8.0",
+    "jest": "^25.1.0",
     "standard-version": "^7.1.0"
   },
   "publishConfig": {
